Fonasa_Inexoos_Flask/requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
Werkzeug==2.2.2
